fix detecting an OpenStack virtual machine under Windows
When using libvirt and kvm a virtual machine has "OpenStack Foundation" as the
System Manufacturer and "OpenStack Nova" as the System Model. This leads to
facter detecting the system as being a physcal machine.

This change adds checking for "OpenStack" in the list of Manufacturers to look
for when detecting whether or not a system is virtual or physical.